move BookImage to app components level
