+++ wbooks/src/app/components/BookImage/index.tsx
