+++ wbooks/src/app/components/BookImage/styles.ts
